remove(ng): extract params
