+++ app/javascript/angular/code/controllers/sessions_map_ctrl.js
@@ -10,7 +10,6 @@ angular
   .module("aircasting")
   .controller("SessionsMapCtrl", [
     "$scope",
-    "params",
     "map",
     sessions,
     "$window",
