+++ app/javascript/angular/code/services/_fixed_sessions.js
@@ -7,9 +7,9 @@ import { sessionsInfoForElm } from "../../../javascript/sessionListUtils";
 import heat_ from "../../../javascript/heat";
 import sensors_ from "../../../javascript/sensors";
 import infoWindow_ from "../../../javascript/infoWindow";
+import params_ from "../../../javascript/params2";
 
-const fixedSessions_ = (heat, infoWindow, sensors) => (
-  params,
+const fixedSessions_ = (heat, infoWindow, params, sensors) => (
   map,
   $rootScope,
   sessionsDownloader,
@@ -230,6 +230,11 @@ export const showClusterInfo = (sensorName, map, infoWindow) => cluster => {
   );
 };
 
-export const fixedSessions = fixedSessions_(heat_, infoWindow_, sensors_);
-export const fixedSessionsTest = (heat, infoWindow, sensors) =>
-  fixedSessions_(heat, infoWindow, sensors);
+export const fixedSessions = fixedSessions_(
+  heat_,
+  infoWindow_,
+  params_,
+  sensors_
+);
+export const fixedSessionsTest = (heat, infoWindow, params, sensors) =>
+  fixedSessions_(heat, infoWindow, params, sensors);
