+++ app/javascript/angular/code/services/fixed_sessions.js
@@ -3,7 +3,6 @@ import { fixedSessions } from "./_fixed_sessions";
 angular
   .module("aircasting")
   .factory("fixedSessions", [
-    "params",
     "map",
     "$rootScope",
     "sessionsDownloader",
