+++ app/javascript/angular/code/services/mobile_sessions.js
@@ -3,7 +3,6 @@ import { mobileSessions } from "./_mobile_sessions";
 angular
   .module("aircasting")
   .factory("mobileSessions", [
-    "params",
     "map",
     "$rootScope",
     "sessionsDownloader",
