+++ app/javascript/angular/code/services/update_crowd_map_layer.js
@@ -5,7 +5,6 @@ angular
   .factory("updateCrowdMapLayer", [
     "map",
     "$http",
-    "params",
     "$window",
     updateCrowdMapLayer
   ]);
