+++ app/javascript/angular/tests/_fixed_sessions_map_ctrl.test.js
@@ -58,9 +58,8 @@ const _SessionsMapCtrl = ({ params }) => {
     subscribe: () => {}
   };
 
-  return SessionsMapCtrlTest(pubsub, _sensors)(
+  return SessionsMapCtrlTest(_params, pubsub, _sensors)(
     _$scope,
-    _params,
     _map,
     sessions,
     $window
