+++ app/javascript/angular/tests/_mobile_sessions.test.js
@@ -354,8 +354,7 @@ const _mobileSessions = ({
   };
   const _heat = { levelName: () => "mid", outsideOfScope: () => false };
 
-  return mobileSessionsTest(_heat, _sensors)(
-    _params,
+  return mobileSessionsTest(_heat, _params, _sensors)(
     _map,
     $rootScope,
     sessionsDownloader,
