+++ app/javascript/angular/tests/_sessions_list_ctrl.test.js
@@ -70,9 +70,8 @@ const _SessionsMapCtrl = ({ sessions, params, pubsub }) => {
     ...pubsub
   };
 
-  return SessionsMapCtrlTest(_pubsub, _sensors)(
+  return SessionsMapCtrlTest(_params, _pubsub, _sensors)(
     _$scope,
-    _params,
     _map,
     _sessions,
     $window
